requirements.txt
flask==3.1.1
limits==5.3.0
loguru==0.7.3
itsdangerous==2.2.0
orjson==3.10.18
waitress==3.0.2
psycopg[binary]==3.2.9
yoyo-migrations==9.0.0
sqlalchemy==2.0.41
pillow==11.2.1
authlib==1.6.0
requests==2.32.4
boto3[crt]==1.35.99
